backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
geoalchemy2==0.15.1
httpx==0.27.0
